+++ donkoProject/src/main/java/classes/user/User.java
@@ -6,6 +6,7 @@ import hash.HashGenerator;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import model.users.usersInsert.InsertNewUserToUsers;
+import model.users.usersSelect.SampleSelectUserFromUsers;
 import model.users.usersSelect.SelectUserDetailFromUsers;
 import model.users.usersSelect.SelectUserDuplicate;
 import model.users.usersSelect.SelectUserFromUsers;
@@ -87,6 +88,11 @@ public class User {
 			return "adminSignin";
 		}
 		return null;
+	};
+	
+	//ログインする
+	public static CustomerUser sampleLogin(CustomerUser customerUser) {
+		return SampleSelectUserFromUsers.sampleSelectUserFromUsers(customerUser);
 	};
 	
 	//ログインする
